Issue#3 Clean up Product Service.
The current implementation of the persistency is very basic.
A lot of improvements are required, but some of these depend on a better version of the ObjestSavingFramework.
There are currently also not test as data is directly written to file and doing file comparisons is pretty nasty with the current state of code.
This part requires some refactering in the future, wich will be done in Issue#14

Closes #3